rename utils.py to graph_spec_utils.py
